backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==1.10.7
psycopg2-binary==2.9.5